MemIO: Applying the IOHandler interface over memory region
io/mem contains the MemIOHandler class, which applies the existing
IOHandler interface to an arbitrary sized region of memory, with no
backing disk. My plan is to use this as the basis for a hashtable, and
then switch over to using a buffered file. Ideally, it should "just
work". We shall see.

Currently there is a major limitation in that the IOHandler API doesn't
provide a good way of checking the available space left in the memory
region. The client must track this on their own. This is simple enough
right now, but the next commit should implement automatic resizing of
the memory buffer, which will make size tracking harder. I suppose, in
princple, you shouldn't need to care about how much space is left in the
case of a resizing buffer, so this might not be such a big limitation
after all--expecially in analogy to files.