Fixed bug in modification flag
Fixed a bug whereby the page would not be marked as "unmodified"
following being flushed. This wouldn't ever be a problem with the
current implementation, as flushing always preceeds the removal of the
page anyway, but it could come up should I later allow flushing of pages
at arbitrary times.