Page::mark_modified changed
I added a mark_modified private method to the buffer manager, which Page
now uses to set the modified flag, rather than directly reading from the
pmeta dictionary itself. We're now in a position where, by making a few
manager methods public, we can drop the friend status of page and
manager, if we wanted to.