Extended Manager Public API
The public API for Buffer::Manager has been extended to include
unpin_page and mark_page_modified, and the friend status of Page has
been dropped. This may be a not-so-great idea, as the client can now
break the synchronization of state between the Page object, and its
corresponding metadata in Manager, but I think this would have become a
problem on if the user handled more than a single page object. I still
need to figure out the best way to handle that situation.